+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 From version 2.6.0, the sections in this file adhere to the [keep a changelog](https://keepachangelog.com/en/1.0.0/) specification.
 ## [Unreleased]
 ### Fixed
+* [#1811](https://github.com/Shopify/shopify-cli/pull/1811): Update theme-check to 1.9.0
 * [#1807](https://github.com/Shopify/shopify-cli/pull/1807): Fix `--live` parameter, it should not imply `--allow-live` in the `theme push` command
 
 ## Version 2.7.2